# src/main/java/com/weihua/model/UserLoginRequest.java
package com.weihua.model;

import lombok.Data;
import javax.validation.constraints.NotBlank;

@Data
public class UserLoginRequest {
    @NotBlank(message = "Username cannot be empty")
    private String username;
    
    @NotBlank(message = "Password cannot be empty")
    private String password;
}