# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  title: Device Management API
  description: API for managing network devices and changing passwords
  version: '1.0.0'
host: localhost:8080
basePath: /api
paths:
...
  /device/monitor/list:
    get:
      summary: List all devices
      description: Get list of all registered devices with optional status filter
      operationId: listDevices
      produces:
        - application/json
      parameters:
        - in: query
          name: status
          description: Filter by device status
          required: false
          type: string
        - in: header
          name: X-Operator-Id
          description: Operator identifier
          required: true
          type: string
      responses:
        '200':
          description: List of devices retrieved
  /device/monitor/{deviceIp}/stats:
    get:
      summary: Get device statistics
      description: Get real-time statistics for a specific device
      operationId: getDeviceStats
      produces:
        - application/json
      parameters:
        - in: path
          name: deviceIp
          description: Device IP address
          required: true
          type: string
        - in: header
          name: X-Operator-Id
          description: Operator identifier
          required: true
          type: string
      responses:
        '200':
          description: Device statistics retrieved
  /device/monitor/{deviceIp}/reboot:
    post:
      summary: Reboot device
      description: Reboot a network device
      operationId: rebootDevice
      produces:
        - application/json
      parameters:
        - in: path
          name: deviceIp
          description: Device IP address
          required: true
          type: string
        - in: header
          name: X-Operator-Id
          description: Operator identifier
          required: true
          type: string
      responses:
        '200':
          description: Reboot initiated successfully
  /device/monitor/alerts:
    get:
      summary: Get device alerts
      description: Get all active alerts for devices
      operationId: getDeviceAlerts
      produces:
        - application/json
      parameters:
        - in: query
          name: level
          description: Filter by alert level
          required: false
          type: string
        - in: header
          name: X-Operator-Id
          description: Operator identifier
          required: true
          type: string
      responses:
        '200':
          description: Alerts retrieved
  /user/login:
    post:
      summary: User login
      description: Authenticate user and get access token
      operationId: userLogin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: request
          description: Login credentials
          required: true
          schema:
            $ref: '#/definitions/UserLoginRequest'
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /user/permissions/{username}:
    get:
      summary: Get user permissions
      description: Get permissions for a specific user
      operationId: getUserPermissions
      produces:
        - application/json
      parameters:
        - in: path
          name: username
          description: Username
          required: true
          type: string
      responses:
        '200':
          description: Permissions retrieved
  /user/validate:
    post:
      summary: Validate token
      description: Validate authentication token
      operationId: validateToken
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Auth-Token
          description: Authentication token
          required: true
          type: string
      responses:
        '200':
          description: Token validation result
...
definitions:
...
  UserLoginRequest:
    type: object
    required:
      - username
      - password
    properties:
      username:
        type: string
        description: Username
        example: "admin"
      password:
        type: string
        description: Password
        example: "admin123"
...
securityDefinitions:
  api_key:
    type: apiKey
    name: X-Auth-Token
    in: header